+++ src/components/AddUser.vue
@@ -9,27 +9,20 @@
 </template>
 
 <script>
-import { ref } from 'vue';
 import UserAlert from './UserAlert.vue';
+import useAlert from '../hooks/alert.js';
 
 export default {
   components: {
     UserAlert,
   },
   setup() {
-    const alertIsVisible = ref(false);
+    const [alertIsVisible, showAlert, hideAlert] = useAlert();
 
-    function showAlert() {
-      alertIsVisible.value = true;
-    }
-    function hideAlert() {
-      alertIsVisible.value = false;
-    }
-    
     return {
       alertIsVisible,
       showAlert,
-      hideAlert
+      hideAlert,
     };
   },
 };
