+++ src/components/DeleteUser.vue
@@ -9,22 +9,15 @@
 </template>
 
 <script>
-import { ref } from 'vue';
 import UserAlert from './UserAlert.vue';
+import useAlert from '../hooks/alert.js';
 
 export default {
   components: {
     UserAlert,
   },
   setup() {
-    const alertIsVisible = ref(false);
-
-    function showAlert() {
-      alertIsVisible.value = true;
-    }
-    function hideAlert() {
-      alertIsVisible.value = false;
-    }
+    const [alertIsVisible, showAlert, hideAlert] = useAlert();
 
     return {
       alertIsVisible,
